refactor: Phase 4 security hardening, hooks consolidation, data access cleanup
Security:
- Add unstable_rethrow to handleServerError — fixes silent redirect failures
- Migrate rate limiter from in-memory Map to @upstash/ratelimit sliding window
- Add path traversal protection to redirectTo validation
- Remove verifyAuthToken re-export from login actions
- Add Cache-Control: no-store on auth redirect responses
- Add frame-ancestors none to Content-Security-Policy
- Add rateLimitedAdminActionClient for rate-limited admin mutations

Hooks:
- Consolidate useActionHandler to /hooks/ (move proven dugsi hook, delete 0-caller root version)
- Update 10 import paths in admin/dugsi components
- Move hook test file to /hooks/__tests__/

Cleanup:
- Remove emoji console.logs from dugsi registration hook
- Remove force-dynamic from admin/dugsi page (was disabling all caching)
- Remove 'use server' from lib/actions/get-students.ts (query functions, not actions)
- Add React.cache() getVerifiedAuth for per-request auth dedup
