+++ app/admin/dugsi/classes/_components/class-form-dialog.tsx
@@ -25,7 +25,7 @@ import {
 } from '@/components/ui/select'
 import { Textarea } from '@/components/ui/textarea'
 
-import { useActionHandler } from '../../_hooks/use-action-handler'
+import { useActionHandler } from '@/hooks/use-action-handler'
 import type { ClassWithDetails } from '../../_types'
 import { createClassAction, updateClassAction } from '../../actions'
