+++ app/admin/dugsi/components/dialogs/add-child-dialog.tsx
@@ -34,7 +34,7 @@ import {
 } from '@/components/ui/select'
 import { Textarea } from '@/components/ui/textarea'
 
-import { useActionHandler } from '../../_hooks/use-action-handler'
+import { useActionHandler } from '@/hooks/use-action-handler'
 import {
   childFormSchema,
   type ChildFormValues,
