+++ app/admin/dugsi/components/dialogs/verify-bank-dialog.tsx
@@ -17,7 +17,7 @@ import {
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 
-import { useActionHandler } from '../../_hooks/use-action-handler'
+import { useActionHandler } from '@/hooks/use-action-handler'
 import { verifyDugsiBankAccount } from '../../actions'
 
 interface VerifyBankDialogProps {
