+++ app/admin/dugsi/components/family-management/detail-tabs/overview-tab.tsx
@@ -9,7 +9,7 @@ import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import { Separator } from '@/components/ui/separator'
 
-import { useActionHandler } from '../../../_hooks/use-action-handler'
+import { useActionHandler } from '@/hooks/use-action-handler'
 import { Family, DugsiRegistration } from '../../../_types'
 import { getOrderedParentData } from '../../../_utils/format'
 import { setPrimaryPayer } from '../../../actions'
