+++ lib/safe-action.ts
@@ -1,4 +1,5 @@
 import { headers } from 'next/headers'
+import { unstable_rethrow } from 'next/navigation'
 import { createSafeActionClient } from 'next-safe-action'
 import { z } from 'zod'
 
@@ -14,6 +15,7 @@ export const actionClient = createSafeActionClient({
     return z.object({ actionName: z.string() })
   },
   async handleServerError(e, utils) {
+    unstable_rethrow(e)
     if (e instanceof ActionError) {
       return e.message
     }
@@ -30,6 +32,24 @@ export const adminActionClient = actionClient.use(async ({ next }) => {
 })
 
 export const rateLimitedActionClient = actionClient.use(
+  async ({ next, metadata }) => {
+    const headersList = await headers()
+    const ip = headersList.get('x-forwarded-for')?.split(',')[0] ?? 'unknown'
+    const key = `${metadata.actionName}:${ip}`
+    const result = await checkRateLimit(key)
+    if (!result.success) {
+      throw new ActionError(
+        'Too many attempts. Please try again later.',
+        ERROR_CODES.RATE_LIMIT_EXCEEDED,
+        undefined,
+        429
+      )
+    }
+    return next()
+  }
+)
+
+export const rateLimitedAdminActionClient = adminActionClient.use(
   async ({ next, metadata }) => {
     const headersList = await headers()
     const ip = headersList.get('x-forwarded-for')?.split(',')[0] ?? 'unknown'
