+++ lib/validations/admin-auth.ts
@@ -5,7 +5,15 @@ export const adminPinSchema = z.object({
     .string()
     .length(4)
     .regex(/^\d{4}$/, 'PIN must be 4 digits'),
-  redirectTo: z.string().startsWith('/admin').optional().default('/admin'),
+  redirectTo: z
+    .string()
+    .startsWith('/admin')
+    .refine(
+      (v) => !v.includes('..') && !v.includes('%2e') && !v.includes('%2E'),
+      'Invalid redirect path'
+    )
+    .optional()
+    .default('/admin'),
 })
 
 export type AdminPinInput = z.infer<typeof adminPinSchema>
